tools/ml_service/requirements.txt
# AI Template Parser - ML Service Dependencies
# 深度学习层次结构分析服务

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0  # PyTorch Image Models

# Graph Neural Networks
torch-geometric>=2.4.0
networkx>=3.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
